files renaming and favicon update
