+++ client/index.html
@@ -2,9 +2,9 @@
 <html lang="fr">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/svg+xml" href="public/Logos/E.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>JS Monorepo</title>
+    <title>Externatech</title>
   </head>
   <body>
     <div id="root"></div>
