+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
-import NavBar from "./components/NavBar/NavBar";
 import Footer from "./components/footer/Footer";
+import NavBar from "./components/navbar/NavBar";
 
 import "./App.css";
