+++ client/src/components/candidateCard/candidateCard.css
