+++ client/src/components/companyCard/companyCard.css
