+++ client/src/components/offerCard/OfferCard.tsx
@@ -1,5 +1,5 @@
 import { Link, useRevalidator } from "react-router-dom";
-import "./offer-card.css";
+import "./offerCard.css";
 import axios from "axios";
 import { useAuth } from "../../services/AuthContext";
