+++ client/src/components/offerCard/offerCard.css
