+++ client/src/pages/companyInformation/companyInformation.css
